Moving Order entity to correct directory
