+++ backend/orders/src/model/Order.kt
@@ -1,6 +1,5 @@
-package challenge
+package challenge.model
 
-import challenge.model.*
 import challenge.service.paymentstrategy.PaymentStrategy
 import java.lang.Exception
 import java.util.*
@@ -17,7 +16,7 @@ data class Order(val customer: Customer, val address: Address) {
     fun addProduct(product: Product, quantity: Int, paymentStrategy: PaymentStrategy) {
         val productAlreadyAdded = items.any { it.product == product }
         if (productAlreadyAdded)
-            throw Exception("The product have already been added. Change the amount if you want more.")
+            throw Exception(DUPLICATED_PRODUCT_EXCEPTION_MESSAGE)
 
         items.add(OrderItem(product, quantity))
         paymentStrategies.add(paymentStrategy)
@@ -28,5 +27,9 @@ data class Order(val customer: Customer, val address: Address) {
     }
 
     override fun toString(): String = "Items: " + this.items + " Total Price: " + totalAmount;
+
+    companion object {
+        const val DUPLICATED_PRODUCT_EXCEPTION_MESSAGE = "Product Free from taxes, according to Constitution Art. 150, VI, d."
+    }
 }
